+++ spec/lib/paypal-api_spec.rb
@@ -70,6 +70,7 @@ RSpec.describe PaypalAPI do
     invoices: PaypalAPI::Invoices,
     orders: PaypalAPI::Orders,
     partner_referrals: PaypalAPI::PartnerReferrals,
+    payment_experience_web_profiles: PaypalAPI::PaymentExperienceWebProfiles,
     payment_tokens: PaypalAPI::PaymentTokens,
     payout_items: PaypalAPI::PayoutItems,
     payouts: PaypalAPI::Payouts,